4.1.14 - Fixes for member coins integer limit

See merge request skycade/skycade-kitpvp!26

+++ src/main/java/net/skycade/kitpvp/commands/CommandRefreshKit.java
@@ -84,7 +84,7 @@ public class CommandRefreshKit extends SkycadeCommand {
         }
 
         KitPvPStats stats = KitPvP.getInstance().getStats(member);
-        int coins = stats.getCoins();
+        long coins = stats.getCoins();
         if (coins < COST) {
             NOT_ENOUGH.msg(member.getPlayer(), "%thing%", "coins");
             return;
