+++ src/main/java/net/skycade/kitpvp/commands/staff/CommandEco.java
@@ -151,7 +151,7 @@ public class CommandEco extends SkycadeCommand {
 
     private void incCoins(KitPvPStats targetStats, CommandSender commandSender, Player target, int amount) {
         targetStats.giveCoins(amount);
-        YOUR_CURRENCY_ADDED.msg(target, "%amount%", Integer.toString(amount), "%currency%", "coins", "%total%", Integer.toString(targetStats.getCoins()));
+        YOUR_CURRENCY_ADDED.msg(target, "%amount%", Integer.toString(amount), "%currency%", "coins", "%total%", Long.toString(targetStats.getCoins()));
         CURRENCY_ADDED.msg(commandSender, "%amount%", Integer.toString(amount), "%currency%", "coins", "%player%", target.getName());
     }
 
@@ -161,7 +161,7 @@ public class CommandEco extends SkycadeCommand {
         else {
             targetStats.takeCoins(targetStats.getCoins() - amount);
             YOUR_CURRENCY_REMOVED.msg(target, "%amount%", Integer.toString(amount), "%currency%", "coins");
-            CURRENCY_REMOVED.msg(commandSender, "%player%", target.getName(), "%amount%", Integer.toString(amount), "%currency%", "coins", "%total%", Integer.toString(targetStats.getCoins()));
+            CURRENCY_REMOVED.msg(commandSender, "%player%", target.getName(), "%amount%", Integer.toString(amount), "%currency%", "coins", "%total%", Long.toString(targetStats.getCoins()));
         }
     }
 }
